If we recover the panic in the Finalize method we can show the recovered error and the stack trace.
As of 9d1dfd0 we discontinued the practice of deferring the Finalize method because we lacked a method of printing out the stack trace and the recovered value.

In this commit we were also able to remove all package-level state (like os.Stdout and the verbose function) by passing these values into the Fixture constructor. This required modifying the generated code, but is a win for testing this framework.