Skipping a fixture prints each skipped test case contained therein.
The Fixture was designed to handle multiple skipped test cases in the case of an entire fixture being marked as skipped but the generated code wasn't invoking that behavior. Now it is.
The generated code was just calling t.Skip(), which is silly because it shut down the goroutine before listing the other skipped test cases.