Fix gunit when GOPATH has multiple directories
The resolveImportPath function attempts to determine the package name
based on the current directory and the GOPATH setting.  However, it
previously assumed that GOPATH had only a single component, while it is
valid for GOPATH to hold multiple directories separated by a colon.

Fix this by making resolveImportPath attempt to replace all components
of GOPATH until a match is found and, if no match is found, fail with a
descriptive error message.  Before this, the only message the user would
get was "cannot import absolute path" which does not point at the
problem.