Display file:line definition location of failed and skipped tests.
We already display file:line info corresponding with the stack of test
failures, but in cases where the failure happened in setup or teardown
we don't have any entry in the stack that points back to the test
definition. This commit fixes #11.

We have never been able to show file:line info corresponding to skipped
tests. This commit fixes #15.

The reason we want to output file:line info is because the GoLand IDE
lacks full support for gunit but what it does do is parse file:line
log entries in the test runner console as links for quick navigation
back to the source code. So, we are taking advantage of a nice built-in
feature to make up for others that are missing. See #12.